ZTRACIM TO

+++ messages.h
@@ -46,7 +46,7 @@ typedef struct {
 } msg_set_compute;
 
 typedef struct {
-    uint8_t cid;  // chunk id
+    double cid;  // chunk id
     double re;    // start of the x-coords (real)
     double im;    // start of the y-coords (imaginary)
     uint8_t n_re; // number of cells in x-coords
@@ -55,7 +55,7 @@ typedef struct {
 } msg_compute;
     
 typedef struct {
-    uint8_t cid;  // chunk id
+    double cid;  // chunk id
     uint8_t i_re; // x-coords
     uint8_t i_im; // y-coords
     uint8_t iter; // number of iterations
